true for overwrite

+++ Commands/ImportWorldCommand.ts
@@ -14,40 +14,45 @@ export class ImportWorldCommand {
         this.manifest = manifest;
     }
 
-    async execute() {
+    async execute(overwrite: boolean = true) {
         new WorldKeyModal(this.app, async (worldKey: string) => {
-            console.log('World key received:', worldKey);
             if (worldKey.length === 10) {
-                try {                
+                try {
                     const response = await requestUrl({
                         url: this.apiUrl + worldKey,
                         method: 'GET'
                     });
-                    console.log('Fetch response:', response);
+
                     if (response.status !== 200) {
                         new Notice('Failed to fetch world data: ' + response.status);
                         return;
                     }
+
                     const data = JSON.parse(response.text);
-                    console.log('World data:', data);
                     const worldData = data['World'];
                     const worldName = worldData ? worldData.name : null;
-                    
+
                     if (!worldName) {
                         new Notice('No valid world data found.');
                         return;
                     }
 
-                    const createTemplatesCommand = new CreateTemplatesCommand(this.app, this.manifest);
-                    await createTemplatesCommand.execute();
-                    
                     const worldFolderPath = normalizePath(`OnlyWorlds/Worlds/${worldName}`);
+                    const fs = this.app.vault.adapter as FileSystemAdapter;
+
+                    // Check if world folder exists, create if not
                     if (!this.app.vault.getAbstractFileByPath(worldFolderPath)) {
                         await this.app.vault.createFolder(worldFolderPath);
                     }
 
-                    await this.generateWorldFile(worldData, worldFolderPath);
-                    await this.generateElementNotes(worldFolderPath, data);
+                    // Generate world file
+                    const worldFilePath = `${worldFolderPath}/World.md`;
+                    if (overwrite || !await fs.exists(worldFilePath)) {
+                        await this.generateWorldFile(worldData, worldFolderPath);
+                    }
+
+                    // Generate element notes
+                    await this.generateElementNotes(worldFolderPath, data, overwrite);
                 } catch (error) {
                     console.error('Error during world import:', error);
                     new Notice('Error fetching world data: ' + error.message);
@@ -57,6 +62,7 @@ export class ImportWorldCommand {
             }
         }).open();
     }
+
 
     async generateWorldFile(worldData: any, worldFolderPath: string) {
         const fs = this.app.vault.adapter as FileSystemAdapter;
@@ -69,37 +75,37 @@ export class ImportWorldCommand {
         console.log(`World file created at: ${worldFilePath}`);
         new Notice(`World file created: ${worldFilePath}`);
     }
-
-    async generateElementNotes(worldFolderPath: string, data: any) {
+    async generateElementNotes(worldFolderPath: string, data: any, overwrite: boolean) {
         const fs = this.app.vault.adapter as FileSystemAdapter;
-        console.log('Starting to generate element notes...');
-    
+        
         for (const category in Category) {
             if (!isNaN(Number(category)) || !data[category]) continue;
-            
+
             const elements = data[category];
             const categoryDirectory = normalizePath(`${worldFolderPath}/${category}`);
+
+            // Check if category folder exists, create if not
             if (!this.app.vault.getAbstractFileByPath(categoryDirectory)) {
                 await this.app.vault.createFolder(categoryDirectory);
             }
-    
-            const templatePath = normalizePath(`${this.app.vault.configDir}/plugins/obsidian-onlyworlds-plugin/Handlebars/${category}Handlebar.md`);
-            const templateText = await fs.read(templatePath);
-            const template = Handlebars.compile(templateText);
-    
+
             for (const element of elements) {
-                let noteContent = template(element);
-                console.log(`Original content for ${element.name}: ${noteContent}`);
-    
-                noteContent = await this.linkifyContent(noteContent, data); // Linkify the content
-                console.log(`Linkified content for ${element.name}: ${noteContent}`);
-    
                 const notePath = `${categoryDirectory}/${element.name}.md`;
-                await fs.write(notePath, noteContent);
-                console.log(`Note created for: ${element.name} at ${notePath}`);
+                
+                if (overwrite || !await fs.exists(notePath)) {
+                    const templatePath = normalizePath(`${this.app.vault.configDir}/plugins/obsidian-onlyworlds-plugin/Handlebars/${category}Handlebar.md`);
+                    const templateText = await fs.read(templatePath);
+                    const template = Handlebars.compile(templateText);
+                    let noteContent = template(element);
+                    
+                    noteContent = await this.linkifyContent(noteContent, data);
+                    
+                    await fs.write(notePath, noteContent);
+                    new Notice(`Note created for: ${element.name}`);
+                }
+          
             }
         }
-        new Notice(`Notes created for all categories.`);
     }
     
     async linkifyContent(noteContent: string, data: any): Promise<string> {
